Voxelization
The program can now voxelize the data set and manage each subdivided voxelized piece via toggle switches. The code needs to be cleaned up so this is essentially a backup.

The code should be partitioned into different class files for better structuring as most of the data is in the ReadData script. Further improvements would also be multi-threading the difference sections of the code as each voxelization set can be on their own thread and the code used is thread safe.