Inclusion of the full project with Updates
The project now includes all of the required assets. Updates include a topographical map as well as the option to change the values for the ranges as well as the colors of the values in the ranges for a voxelized mesh set.

Work needed:
The color changes should reflect the shell as well.
The topographical map is not fitted towards a lat/lon as the voxelized data sets are.
The color ranges should be checked for errors on change so that overlapping color ranges do not occour.