Updated Version and improvements
Version updated to Unity beta 2018.1.0b9

Added changes:

Topography mapped to elevation scale changes on the slider.

Show/Hide Topography toggle added.

Fixes:

Fixed error in removal of material color ranges.

Known issues:
Adjusted topography mesh position on elevation scale change is not correct as a parameter added to 5.0f makes it look correct for the current elevation value ranges clamped.

Planned fixes:

The topography is not displayed in a correct lat/lon view as the underlaying data is presented as. The goal is to have both a linear box view( which works witht he current topography model) and a latlon representation.

The bounding box view needs to be updated as it is not currently encompassing the underlaying data as it should be.

Lon / Lat coordinates need to be placed over the mesh to give the user a better visual aid of the data range being viewed.