Topography heightMap color update
The topography color display was formerly determined via color picker. This was not a very good implementation as the entire topology was shaded in that color. The new implementation takes in the elevation values and maps a color to the vertices according to the reference on the imported colormap picture. The elevation range I chose was from -6000 to +6000 which were the default values used in MatLab. I also modified the custom vertex shader to allow for an alpha channel.  newly created Alpha slider has been implemented below the elveation scale slider. Changing this slider directly calls the material shader attached to the topography and changes the alpha there. It is a quick solution.