Fixed topography position
Fixed the positions of the topography, shell structure and voxelized meshes.  Original positions where to be centered in the world coordinates with the camera orbiting the center position. The positions now stay where the obejcts are constructed and parented under a root object which is then offset so that the camera can rotate around the root object. This allows for larger topography data to be loaded that does not fit the underlying volue data and still be in correct position.